requirements.txt
pandas==1.2.4
matplotlib==3.3.4
plotly==4.14.3
streamlit_drawable_canvas==0.8.0
streamlit==0.82.0
numpy==1.19.5
torchvision==0.9.1
torch==1.8.1
Pillow==8.2.0
